file renames and import updates
